+++ include/layered_hardware_gazebo/velocity_mode.hpp
@@ -42,7 +42,11 @@ public:
     data->effort = joint_->GetForce(0);
   }
 
-  virtual void write(const ti::RawJointData &data) { vel_cmd_ = data.velocity_cmd; }
+  virtual void write(const ti::RawJointData &data) {
+    if (!boost::math::isnan(data.velocity_cmd)) {
+      vel_cmd_ = data.velocity_cmd;
+    }
+  }
 
   virtual void update(const ros::Time &time, const ros::Duration &period) {
     if (!boost::math::isnan(vel_cmd_)) {